+++ Editor/Utility/YarnProjectUtility.cs
@@ -43,24 +43,81 @@ namespace Yarn.Unity.Editor
 
             // Create the program
             YarnEditorUtility.CreateYarnAsset(destinationPath);
-
+            
             AssetDatabase.ImportAsset(destinationPath);
             AssetDatabase.SaveAssets();
 
-            var programImporter = AssetImporter.GetAtPath(destinationPath) as YarnProjectImporter;
-            programImporter.sourceScripts.Add(AssetDatabase.LoadAssetAtPath<TextAsset>(path));
-
-            EditorUtility.SetDirty(programImporter);
-
-            // Reimport the program to make it generate its default string
-            // table, if needed
-            programImporter.SaveAndReimport();
+            AssignScriptToProject(path, destinationPath);
 
             return destinationPath;
-
-
         }
 
+        /// <summary>
+        /// Assign a .yarn <see cref="TextAsset"/> file found at <paramref
+        /// name="sourcePath"/> to the <see cref="YarnProjectImporter"/>
+        /// found at <paramref name="projectPath"/>.
+        /// </summary>
+        /// <param name="projectPath">The path to the .yarnproject
+        /// asset.</param>
+        /// <param name="sourcePath">The path to the source .yarn asset
+        /// that should be added to the Yarn Project.</param>
+        internal static void AssignScriptToProject(string sourcePath, string projectPath) {
+            var newSourceScript = AssetDatabase.LoadAssetAtPath<TextAsset>(sourcePath);
+            var programImporter = AssetImporter.GetAtPath(projectPath) as YarnProjectImporter;
+            var scriptImporter = AssetImporter.GetAtPath(sourcePath) as YarnImporter;
+
+            if (newSourceScript == null) {
+                throw new FileNotFoundException($"{nameof(sourcePath)} couldn't be loaded as a {nameof(TextAsset)}.");
+            }
+
+            if (projectPath != null && programImporter == null) {
+                throw new System.ArgumentException($"{nameof(projectPath)} is not a Yarn Project.");
+            }
+
+            if (scriptImporter == null) {
+                throw new System.ArgumentException($"{nameof(projectPath)} is not a Yarn script.");
+            }
+
+            AssignScriptToProject(newSourceScript, scriptImporter, programImporter);
+        }
+
+        /// <summary>
+        /// Assign a .yarn <see cref="TextAsset"/> file <paramref
+        /// name="newSourceScript"/> to the <see
+        /// cref="YarnProjectImporter"/> <paramref
+        /// name="projectImporter"/>.
+        /// </summary>
+        /// <remarks>If <paramref name="projectImporter"/> is <see
+        /// langword="null"/>, this script will be removed from its current
+        /// project (if any) and not added to another.</remarks>
+        /// <param name="newSourceScript">The script that should be
+        /// assigned to the Yarn Project.</param>
+        /// <param name="scriptImporter">The importer for <paramref
+        /// name="newSourceScript"/>.</param>
+        /// <param name="projectImporter">The importer for the project that
+        /// newSourceScript should be made a part of, or null.</param>
+        internal static void AssignScriptToProject(TextAsset newSourceScript, YarnImporter scriptImporter, YarnProjectImporter projectImporter) {
+
+            if (scriptImporter.DestinationProject != null) {
+                var existingProjectImporter = AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(scriptImporter.DestinationProject)) as YarnProjectImporter;
+
+                existingProjectImporter.sourceScripts.Remove(newSourceScript);
+
+                EditorUtility.SetDirty(existingProjectImporter);
+                
+                existingProjectImporter.SaveAndReimport();
+            }
+            
+            if (projectImporter != null) {
+                projectImporter.sourceScripts.Add(newSourceScript);
+                EditorUtility.SetDirty(projectImporter);
+
+                // Reimport the program to make it generate its default string
+                // table, if needed
+                projectImporter.SaveAndReimport();
+            }
+
+        }
 
         /// <summary>
         /// Updates every localization .CSV file associated with this
@@ -230,7 +287,7 @@ namespace Yarn.Unity.Editor
                 Compare.By<string>((fileName, lineID) =>
                 {
                     var lineIDWithoutPrefix = lineID.Replace("line:", "");
-                    return fileName.Contains(lineIDWithoutPrefix);
+                    return Path.GetFileNameWithoutExtension(fileName).Equals(lineIDWithoutPrefix);
                 })
                 )
                 // Discard any pair where no asset was found
@@ -365,10 +422,10 @@ namespace Yarn.Unity.Editor
 
         internal static void AddLineTagsToFilesInYarnProject(YarnProjectImporter importer)
         {
-            // First, gather all existing line tags, so that we don't
-            // accidentally overwrite an existing one. Do this by finding
-            // all yarn scripts in all yarn projects, and get the string
-            // tags inside them.
+            // First, gather all existing line tags across ALL yarn
+            // projects, so that we don't accidentally overwrite an
+            // existing one. Do this by finding all yarn scripts in all
+            // yarn projects, and get the string tags inside them.
 
             var allYarnFiles =
                 // get all yarn projects across the entire project
@@ -402,21 +459,22 @@ namespace Yarn.Unity.Editor
             // if a file contains a parse error.
             var allExistingTags = allYarnFiles.SelectMany(path =>
             {
-                try
-                {
-                    // Compile this script in strings-only mode to get
-                    // string entries
-                    var compilationJob = Yarn.Compiler.CompilationJob.CreateFromFiles(path);
-                    compilationJob.CompilationType = Yarn.Compiler.CompilationJob.Type.StringsOnly;
+                // Compile this script in strings-only mode to get
+                // string entries
+                var compilationJob = Yarn.Compiler.CompilationJob.CreateFromFiles(path);
+                compilationJob.CompilationType = Yarn.Compiler.CompilationJob.Type.StringsOnly;
 
-                    var result = Yarn.Compiler.Compiler.Compile(compilationJob);
-                    return result.StringTable.Where(i => i.Value.isImplicitTag == false).Select(i => i.Key);
-                }
-                catch (Yarn.Compiler.CompilerException e)
-                {
-                    Debug.LogWarning($"Can't check for existing line tags in {path}, because a compiler exception was thrown: {e}");
+                var result = Yarn.Compiler.Compiler.Compile(compilationJob);
+
+                bool containsErrors = result.Diagnostics
+                    .Any(d => d.Severity == Compiler.Diagnostic.DiagnosticSeverity.Error);
+
+                if (containsErrors) {
+                    Debug.LogWarning($"Can't check for existing line tags in {path} because it contains errors.");
                     return new string[] { };
                 }
+                
+                return result.StringTable.Where(i => i.Value.isImplicitTag == false).Select(i => i.Key);
             }).ToList(); // immediately execute this query so we can determine timing information
 
 #if YARNSPINNER_DEBUG
@@ -511,6 +569,46 @@ namespace Yarn.Unity.Editor
             File.WriteAllText(destination, outputCSV);
 
             return true;
+        }
+
+        internal static void ConvertImplicitVariableDeclarationsToExplicit(YarnProjectImporter yarnProjectImporter) {
+            var allFilePaths = yarnProjectImporter.sourceScripts.Select(textAsset => AssetDatabase.GetAssetPath(textAsset));
+
+            var library = new Library();
+            YarnProject.AddYarnFunctionMethodsToLibrary(library);
+
+            var explicitDeclarationsCompilerJob = Compiler.CompilationJob.CreateFromFiles(AssetDatabase.GetAssetPath(yarnProjectImporter));
+
+            Compiler.CompilationResult explicitResult;
+
+            try {
+                explicitResult = Compiler.Compiler.Compile(explicitDeclarationsCompilerJob);
+            } catch (System.Exception e) {
+                Debug.LogError($"Compile error: {e}");
+                return;
+            }
+
+
+            var implicitDeclarationsCompilerJob = Compiler.CompilationJob.CreateFromFiles(allFilePaths, library);
+            implicitDeclarationsCompilerJob.CompilationType = Compiler.CompilationJob.Type.DeclarationsOnly;
+            implicitDeclarationsCompilerJob.VariableDeclarations = explicitResult.Declarations;
+
+            Compiler.CompilationResult implicitResult;
+
+            try {
+                implicitResult = Compiler.Compiler.Compile(implicitDeclarationsCompilerJob);
+            } catch (System.Exception e) {
+                Debug.LogError($"Compile error: {e}");
+                return;
+            }
+
+            var implicitDeclarations = implicitResult.Declarations.Where(d => !(d.Type is Yarn.FunctionType) && d.IsImplicit);
+
+            var output = Yarn.Compiler.Utility.GenerateYarnFileWithDeclarations(explicitResult.Declarations.Concat(implicitDeclarations), "Program");
+
+            File.WriteAllText(yarnProjectImporter.assetPath, output, System.Text.Encoding.UTF8);
+            AssetDatabase.ImportAsset(yarnProjectImporter.assetPath);
+
         }